@@ -3,34 +3,34 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@mdi/js": "^6.4.95",
-    "@mdi/react": "^1.5.0",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.34",
-    "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.20",
-    "@types/react-router-dom": "^5.1.7",
-    "axios": "^0.24.0",
-    "external-svg-loader": "^1.3.4",
-    "http-proxy-middleware": "^2.0.1",
-    "jwt-decode": "^3.1.2",
-    "react": "^17.0.2",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^5.2.0",
+    "@mdi/js": "^7.4.47",
+    "@mdi/react": "^1.6.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.10.6",
+    "@types/react": "^18.3.18",
+    "@types/react-beautiful-dnd": "^13.1.8",
+    "@types/react-dom": "^18.3.5",
+    "@types/react-redux": "^7.1.33",
+    "@types/react-router-dom": "^5.3.3",
+    "axios": "^1.7.9",
+    "external-svg-loader": "^1.6.9",
+    "http-proxy-middleware": "^3.0.3",
+    "jwt-decode": "^4.0.0",
+    "react": "^18.3.1",
+    "react-beautiful-dnd": "^13.1.1",
+    "react-dom": "^18.3.1",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^5.3.4",
     "react-scripts": "^5.0.1",
-    "redux": "^4.1.2",
+    "redux": "^5.0.1",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.4.0",
-    "skycons-ts": "^0.2.0",
-    "typescript": "^4.4.4",
-    "web-vitals": "^2.1.2"
+    "redux-thunk": "^3.1.0",
+    "skycons-ts": "^1.0.0",
+    "typescript": "^5.7.2",
+    "web-vitals": "^4.2.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -57,6 +57,6 @@
     ]
   },
   "devDependencies": {
-    "prettier": "^2.4.1"
+    "prettier": "^3.4.2"
   }
 }
