@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import thunk from 'redux-thunk';
+import { thunk } from 'redux-thunk';
 import { reducers } from './reducers';
 
 export const store = createStore(
